streamlit/requirements.txt
streamlit
tensorflow
tensorflow-datasets
matplotlib
pillow
